+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ## Master
 
+- SERVER: Updated .COM, .NET TLD definitions.
+
 - CHANGED: Updated whois.nic.tv parser to the new response format (GH-283).
 
 - CHANGED: Updated Verisign parsers, extract common parser.